iCloush_Backend_V1/requirements.txt
# ═══════════════════════════════════════════════════
# iCloush 智慧工厂后端 — Python 依赖
# ═══════════════════════════════════════════════════

# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# 数据库
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# 配置
pydantic-settings==2.5.2
python-dotenv==1.0.1

# 认证
PyJWT==2.9.0

# HTTP 客户端
httpx==0.27.2

# Redis（纯 Python 版，避免 hiredis C 扩展在 slim 镜像上 Segfault）
redis==5.1.1

# 腾讯云 COS STS（可选）
# qcloud-python-sts==3.1.6

# 工具
orjson==3.10.7
